requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
azure-common==1.1.28
azure-core==1.31.0
azure-search-documents==11.5.1
azure-storage-blob==12.23.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isodate==0.7.0
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.17
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langsmith==0.1.132
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.51.2
orjson==3.10.7
packaging==23.2
propcache==0.2.0
psycopg2==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pypdf==4.3.1
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tenacity==8.5.0
tiktoken==0.8.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
yarl==1.14.0
